+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
-### Unreleased
+### [v0.13](https://github.com/realityforge/vecmath/tree/v0.13) (2021-11-16) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.12...v0.13)
+
+Changes in this release:
 
 * Update the `org.realityforge.grim` artifacts to version `0.06`.
 * Add `Vector3d.mult(Matrix4d)` method.